CLEANUP: Move event callback functions to mc_util files
