+++ memcached.h
@@ -250,12 +250,6 @@ struct settings {
     } extensions;
 };
 
-struct engine_event_handler {
-    EVENT_CALLBACK cb;
-    const void *cb_data;
-    struct engine_event_handler *next;
-};
-
 extern struct settings settings;
 extern EXTENSION_LOGGER_DESCRIPTOR *mc_logger;
